requirements.txt
streamlit
langchain
langgraph
langchain-openai
langchain-community
openai
pypdf
python-dotenv
pdfkit
langchain-openai
langchain-anthropic
langchain-google-genai
anthropic>=0.25
google-generativeai>=0.5
newspaper3k
lxml[html_clean]
